refactor: move check and checkmate logic to board utils, add helper function to get king cell
